Modification de la page d'accueil, déplacement de la page comment_ca_marche/index.php; Css
