--- Controllers/ArticlesController.php
@@ -1,10 +1,0 @@
-<?php
-namespace App\Controllers;
-
-class ArticlesController extends Controller
-{
-    public function index()
-    {
-        
-    }
-}
